+++ pyres/__init__.py
@@ -1,12 +1,14 @@
-__version__ = '0.5.0'
+__version__ = '0.7.5.1'
 
 from redis import Redis
 import pyres.json_parser as json
 
+import time, datetime
+
 import logging
-import types
 
 def my_import(name):
+    """Helper function for walking import calls when searching for classes by string names."""
     mod = __import__(name)    
     components = name.split('.')    
     for comp in components[1:]:        
@@ -14,20 +16,19 @@ def my_import(name):
     return mod
 
 def safe_str_to_class(s):
+    """Helper function to map string class names to module classes."""
     lst = s.split(".")
     klass = lst[-1]
     mod_list = lst[:-1]
     module = ".".join(mod_list)
-    try:
-        mod = my_import(module)
-        if hasattr(mod, klass):
-            return getattr(mod, klass)
-        else:
-            return None
-    except ImportError:
-        return None
+    mod = my_import(module)
+    if hasattr(mod, klass):
+        return getattr(mod, klass)
+    else:
+        raise ImportError('')
 
 def str_to_class(s):
+    """Alternate helper function to map string class names to module classes."""
     lst = s.split(".")
     klass = lst[-1]
     mod_list = lst[:-1]
@@ -42,8 +43,22 @@ def str_to_class(s):
         return None
 
 class ResQ(object):
-    """ResQ class which defines the Queue object to enqueue jobs into various
-    queues.
+    """The ResQ class defines the Redis server object to which we will
+    enqueue jobs into various queues.
+    
+    The ``__init__`` takes these keyword arguments:
+    
+        ``server`` -- IP address and port of the Redis server to which you want to connect. Default is `localhost:6379`.
+    
+        ``password`` -- The password, if required, of your Redis server. Default is "None".
+    
+        ``timeout`` -- The timeout keyword is in the signature, but is unused. Default is "None".
+    
+        ``retry_connection`` -- This keyword is in the signature but is deprecated. Default is "True".
+    
+    
+    Both ``timeout`` and ``retry_connection`` will be removed as the python-redis client
+    no longer uses them. 
     
     Example usage::
 
@@ -55,13 +70,11 @@ class ResQ(object):
 
         >>>> r.enqueue(SomeClass, args)
 
-    SomeClass can be any python class with *perform* method and a *queue* 
+    SomeClass can be any python class with a *perform* method and a *queue* 
     attribute on it.
+    
     """
-    def __init__(self, server="localhost:6379", password=None, 
-                 timeout=None, retry_connection=True):
-        self.timeout = timeout
-        #self.retry_connection = retry_connection
+    def __init__(self, server="localhost:6379", password=None):
         self.redis = server
         if password:
             self.redis.auth(password)
@@ -113,9 +126,9 @@ class ResQ(object):
     redis = property(_get_redis, _set_redis)
 
     def enqueue(self, klass, *args):
-        """
-        Enqueue a job into a specific queue. Make sure the class you are passing
+        """Enqueue a job into a specific queue. Make sure the class you are passing
         has **queue** attribute and a **perform** method on it.
+        
         """
         queue = getattr(klass,'queue', None)
         if queue:
@@ -129,8 +142,11 @@ class ResQ(object):
         else:
             logging.warning("unable to enqueue job with class %s" % str(klass))
 
-    def enqueue_from_string(self, klass_as_string, queue, *args):
-        self.push(queue, {'class':klass_as_string,'args':args})
+    def enqueue_from_string(self, klass_as_string, queue, *args, **kwargs):
+        payload = {'class':klass_as_string, 'queue': queue, 'args':args}
+        if 'first_attempt' in kwargs:
+            payload['first_attempt'] = kwargs['first_attempt']
+        self.push(queue, payload)
         logging.info("enqueued '%s' job" % klass_as_string)
         if args:
             logging.debug("job arguments: %s" % str(args))
@@ -141,9 +157,9 @@ class ResQ(object):
         return self.redis.smembers("resque:queues") or []
     
     def info(self):
-        """
-        Returns a dictionary of the current status of the pending jobs, 
+        """Returns a dictionary of the current status of the pending jobs, 
         processed, no. of queues, no. of workers, no. of failed jobs.
+        
         """
         pending = 0
         for q in self.queues():
@@ -183,10 +199,58 @@ class ResQ(object):
         del self.redis['resque:queue:%s' % queue]
     
     def close(self):
-        """
-        close the underlying redis connection
+        """Close the underlying redis connection.
+        
         """
         self.redis.disconnect()
+    
+    def enqueue_at(self, datetime, klass, *args, **kwargs):
+        class_name = '%s.%s' % (klass.__module__, klass.__name__)
+        logging.info("enqueued '%s' job for execution at %s" % (class_name, datetime))
+        if args:
+            logging.debug("job arguments are: %s" % str(args))
+        payload = {'class':class_name, 'queue': klass.queue, 'args':args}
+        if 'first_attempt' in kwargs:
+            payload['first_attempt'] = kwargs['first_attempt']
+        self.delayed_push(datetime, payload)
+    
+    def delayed_push(self, datetime, item):
+        key = int(time.mktime(datetime.timetuple()))
+        self.redis.rpush('resque:delayed:%s' % key, ResQ.encode(item))
+        self.redis.zadd('resque:delayed_queue_schedule', key, key)
+    
+    def delayed_queue_peek(self, start, count):
+        return [int(item) for item in self.redis.zrange('resque:delayed_queue_schedule', start, start+count) or []]
+    
+    def delayed_timestamp_peek(self, timestamp, start, count):
+        return self.list_range('resque:delayed:%s' % timestamp, start, count)
+        
+    def delayed_queue_schedule_size(self):
+        return self.redis.zcard('resque:delayed_queue_schedule')
+    
+    def delayed_timestamp_size(self, timestamp):
+        #key = int(time.mktime(timestamp.timetuple()))
+        return self.redis.llen("resque:delayed:%s" % timestamp)
+    
+    def next_delayed_timestamp(self):
+        key = int(time.mktime(ResQ._current_time().timetuple()))
+        array = self.redis.zrangebyscore('resque:delayed_queue_schedule', '-inf', key)
+        timestamp = None
+        if array:
+            timestamp = array[0]
+        return timestamp
+    
+    def next_item_for_timestamp(self, timestamp):
+        #key = int(time.mktime(timestamp.timetuple()))
+        key = "resque:delayed:%s" % timestamp
+        ret = self.redis.lpop(key)
+        item = None
+        if ret:
+            item = ResQ.decode(ret)
+        if self.redis.llen(key) == 0:
+            self.redis.delete(key)
+            self.redis.zrem('resque:delayed_queue_schedule', timestamp)
+        return item
     
     @classmethod
     def encode(cls, item):
@@ -207,10 +271,14 @@ class ResQ(object):
             class_name = '%s.%s' % (klass.__module__, klass.__name__)
             _self.push(queue, {'class':class_name,'args':args})
 
+    @staticmethod
+    def _current_time():
+        return datetime.datetime.now()
+
 
 class Stat(object):
-    """
-    A Stat class which shows the current status of the queue.
+    """A Stat class which shows the current status of the queue.
+    
     """
     def __init__(self, name, resq):
         self.name = name