+++ pyres/failure/redis.py
@@ -1,12 +1,16 @@
-import datetime
+import datetime, time
 from base import BaseBackend
 from pyres import ResQ
+
 class RedisBackend(BaseBackend):
+    """Extends the ``BaseBackend`` to provide a Redis backend for failed jobs."""
+    
     def save(self, resq=None):
+        """Saves the failed Job into a "failed" Redis queue preserving all its original enqueud info."""
         if not resq:
             resq = ResQ()
         data = {
-            'failed_at' : str(datetime.datetime.now()),
+            'failed_at' : int(time.mktime(datetime.datetime.now().timetuple())),
             'payload'   : self._payload,
             'error'     : self._parse_message(self._exception),
             'backtrace' : self._parse_traceback(self._traceback),