+++ pyres/worker.py
@@ -3,19 +3,19 @@ from pyres.job import Job
 from pyres import ResQ, Stat
 import logging
 import signal
-import datetime
+import datetime, time
 import os, sys
 import time
 import json_parser as json
 import commands
 
 class Worker(object):
-    """
-    Defines a worker. The *pyres_worker* script instantiates this Worker class and
-    pass a comma seperate list of queues to listen on.::
+    """Defines a worker. The ``pyres_worker`` script instantiates this Worker class and
+    passes a comma-separated list of queues to listen on.::
     
        >>> from pyres.worker import Worker
        >>> Worker.run([queue1, queue2], server="localhost:6379")
+       
     """
     def __init__(self, queues=[], server="localhost:6379", password=None):
         self.queues = queues
@@ -33,7 +33,7 @@ class Worker(object):
         
     
     def validate_queues(self):
-        "Checks if a worker is given atleast one queue to work on."
+        """Checks if a worker is given at least one queue to work on."""
         if not self.queues:
             raise NoQueueError("Please give each worker at least one queue.")
     
@@ -43,18 +43,19 @@ class Worker(object):
         self.started = datetime.datetime.now()
         
     
-    def _set_started(self, time):
-        if time:
-            self.resq.redis.set("resque:worker:%s:started" % self, time.strftime('%Y-%m-%d %H:%M:%S'))
+    def _set_started(self, dt):
+        if dt:
+            key = int(time.mktime(dt.timetuple()))
+            self.resq.redis.set("resque:worker:%s:started" % self, key)
         else:
             self.resq.redis.delete("resque:worker:%s:started" % self)
             
     def _get_started(self):
         datestring = self.resq.redis.get("resque:worker:%s:started" % self)
-        ds = None
-        if datestring:
-            ds = datetime.datetime.strptime(datestring, '%Y-%m-%d %H:%M:%S')
-        return ds
+        #ds = None
+        #if datestring:
+        #    ds = datetime.datetime.strptime(datestring, '%Y-%m-%d %H:%M:%S')
+        return datestring
     
     started = property(_get_started, _set_started)
     
@@ -67,7 +68,7 @@ class Worker(object):
     def prune_dead_workers(self):
         all_workers = Worker.all(self.resq)
         known_workers = self.worker_pids()
-        
+
         logging.debug("all workers: '%s'" % all_workers)
         logging.debug("known workers: '%s'" % known_workers)
         logging.debug("hostname: '%s'" % self.hostname)
@@ -110,16 +111,17 @@ class Worker(object):
         return '%s:%s:%s' % (self.hostname, self.pid, ','.join(self.queues))
          
     def work(self, interval=5):
-        """Invoked by run() method. work() listens on a list of queues and sleeps
-        for *interval* time. 
+        """Invoked by ``run`` method. ``work`` listens on a list of queues and sleeps
+        for ``interval`` time. 
         
-        default  --  5 secs
+        ``interval`` -- Number of seconds the worker will wait until processing the next job. Default is "5".
         
         Whenever a worker finds a job on the queue it first calls ``reserve`` on
-        that job to make sure other worker won't run it, then *Forks* itself to 
+        that job to make sure another worker won't run it, then *forks* itself to 
         work on that job.
         
-        Finally process() method actually processes the job.
+        Finally, the ``process`` method actually processes the job by eventually calling the Job instance's ``perform`` method.
+        
         """
         self.startup()
         while True:
@@ -157,7 +159,7 @@ class Worker(object):
             job = self.reserve()
         try:
             self.working_on(job)
-            job.perform()
+            return job.perform()
         except Exception, e:
             exceptionType, exceptionValue, exceptionTraceback = sys.exc_info()
             logging.error("%s failed: %s" % (job, e))
@@ -181,7 +183,7 @@ class Worker(object):
         logging.debug('marking as working on')
         data = {
             'queue': job._queue,
-            'run_at': str(datetime.datetime.now()),
+            'run_at': int(time.mktime(datetime.datetime.now().timetuple())),
             'payload': job._payload
         }
         data = json.dumps(data)